@@ -47,7 +47,7 @@ class _ArticlesViewState extends State<ArticlesView> {
             padding: const EdgeInsets.all(8.0),
             child: ListView.builder(
                 physics: const AlwaysScrollableScrollPhysics(),
-                itemCount: model.articlesModel.results.length,
+                itemCount: model.articlesDto.results.length,
                 itemBuilder: (context, index) {
                   return ArticlesCard(model: model, index: index);
                 }),
